Move common code to a separate file
